Add canvas reducer
