+++ packages/noya-state/src/reducers/applicationReducer.ts
@@ -1,32 +1,23 @@
 import Sketch from '@sketch-hq/sketch-file-format-ts';
 import produce from 'immer';
 import { WritableDraft } from 'immer/dist/internal';
-import { AffineTransform, normalizeRect } from 'noya-geometry';
-import { Primitives, SimpleTextDecoration, uuid } from 'noya-renderer';
-import { resizeRect } from 'noya-renderer/src/primitives';
+import { SimpleTextDecoration, uuid } from 'noya-renderer';
 import { SketchFile } from 'noya-sketch-file';
 import { IndexPath } from 'tree-visit';
 import * as Layers from '../layers';
-import * as Models from '../models';
 import {
-  EncodedPageMetadata,
   findPageLayerIndexPaths,
-  getBoundingRect,
   getCurrentComponentsTab,
   getCurrentPage,
   getCurrentPageIndex,
-  getCurrentPageMetadata,
   getCurrentTab,
   getSelectedLayerIndexPaths,
-  getSelectedLayerIndexPathsExcludingDescendants,
 } from '../selectors/selectors';
-import { Point } from '../types';
 import { SelectionType, updateSelection } from '../utils/selection';
 import { AlignmentAction, alignmentReducer } from './alignmentReducer';
+import { CanvasAction, canvasReducer } from './canvasReducer';
 import {
-  CompassDirection,
   createInitialInteractionState,
-  InteractionAction,
   interactionReducer,
   InteractionState,
 } from './interactionReducer';
@@ -66,26 +57,12 @@ export type ApplicationState = {
 export type Action =
   | [type: 'setTab', value: WorkspaceTab]
   | PageAction
-  | [
-      type: 'insertArtboard',
-      details: { name: string; width: number; height: number },
-    ]
-  | [type: 'addDrawnLayer']
+  | CanvasAction
   | [type: 'deleteLayer', layerId: string | string[]]
   | [
       type: 'selectLayer',
       layerId: string | string[] | undefined,
       selectionType?: SelectionType,
-    ]
-  | [
-      type: 'interaction',
-      // Some actions may need to be augmented by additional state before
-      // being passed to nested reducers (e.g. `maybeScale` takes a snapshot
-      // of the current page). Maybe there's a better way? This still seems
-      // better than moving the whole reducer up into the parent.
-      action:
-        | Exclude<InteractionAction, ['maybeScale', ...any[]]>
-        | [type: 'maybeScale', origin: Point, direction: CompassDirection],
     ]
   | StyleAction
   | StringAttributeAction
@@ -119,57 +96,15 @@ export function applicationReducer(
     case 'movePage': {
       return pageReducer(state, action);
     }
-    case 'insertArtboard': {
-      const [, { name, width, height }] = action;
-      const pageIndex = getCurrentPageIndex(state);
-      const { scrollOrigin } = getCurrentPageMetadata(state);
-
-      return produce(state, (draft) => {
-        let layer = produce(Models.artboard, (layer) => {
-          layer.do_objectID = uuid();
-          layer.name = name;
-          layer.frame = {
-            _class: 'rect',
-            constrainProportions: false,
-            // TODO: Figure out positioning based on other artboards.
-            // Also, don't hardcode sidebar width.
-            x: -scrollOrigin.x + 100,
-            y: -scrollOrigin.y + 100,
-            width,
-            height,
-          };
-        });
-
-        draft.sketch.pages[pageIndex].layers.push(layer);
-        draft.interactionState = interactionReducer(draft.interactionState, [
-          'reset',
-        ]);
-        draft.selectedObjects = [layer.do_objectID];
-      });
-    }
+    case 'insertArtboard':
+    case 'addDrawnLayer':
+    case 'interaction':
+      return canvasReducer(state, action);
     case 'setLayerVisible':
     case 'setExpandedInLayerList':
     case 'setFixedRadius':
     case 'setLayerRotation':
       return layerPropertyReducer(state, action);
-    case 'addDrawnLayer': {
-      const pageIndex = getCurrentPageIndex(state);
-
-      return produce(state, (draft) => {
-        if (draft.interactionState.type !== 'drawing') return;
-
-        const layer = draft.interactionState.value;
-
-        if (layer.frame.width > 0 && layer.frame.height > 0) {
-          draft.sketch.pages[pageIndex].layers.push(layer);
-          draft.selectedObjects = [layer.do_objectID];
-        }
-
-        draft.interactionState = interactionReducer(draft.interactionState, [
-          'reset',
-        ]);
-      });
-    }
     case 'deleteLayer': {
       const [, id] = action;
 
@@ -203,7 +138,6 @@ export function applicationReducer(
         updateSelection(draft.selectedObjects, id, selectionType);
       });
     }
-
     case 'distributeLayers':
     case 'alignLayers': {
       return alignmentReducer(state, action);
@@ -290,161 +224,6 @@ export function applicationReducer(
           });
         });
       }
-    }
-    case 'interaction': {
-      const page = getCurrentPage(state);
-      const currentPageId = page.do_objectID;
-      const pageIndex = getCurrentPageIndex(state);
-      const layerIndexPaths = getSelectedLayerIndexPathsExcludingDescendants(
-        state,
-      );
-      const layerIds = layerIndexPaths.map(
-        (indexPath) => Layers.access(page, indexPath).do_objectID,
-      );
-
-      const interactionState = interactionReducer(
-        state.interactionState,
-        action[1][0] === 'maybeScale' ? [...action[1], page] : action[1],
-      );
-      return produce(state, (draft) => {
-        draft.interactionState = interactionState;
-
-        switch (interactionState.type) {
-          case 'moving': {
-            const { previous, next } = interactionState;
-
-            const delta = {
-              x: next.x - previous.x,
-              y: next.y - previous.y,
-            };
-
-            accessPageLayers(draft, pageIndex, layerIndexPaths).forEach(
-              (layer) => {
-                layer.frame.x += delta.x;
-                layer.frame.y += delta.y;
-              },
-            );
-
-            break;
-          }
-          case 'scaling': {
-            const {
-              origin,
-              current,
-              pageSnapshot,
-              direction,
-            } = interactionState;
-
-            const originalBoundingRect = getBoundingRect(
-              pageSnapshot,
-              AffineTransform.identity,
-              layerIds,
-            )!;
-
-            const newBoundingRect = resizeRect(
-              originalBoundingRect,
-              {
-                x: current.x - origin.x,
-                y: current.y - origin.y,
-              },
-              direction,
-            );
-
-            const originalTransform = AffineTransform.multiply(
-              AffineTransform.translation(
-                originalBoundingRect.x,
-                originalBoundingRect.y,
-              ),
-              AffineTransform.scale(
-                originalBoundingRect.width,
-                originalBoundingRect.height,
-              ),
-            ).invert();
-
-            const newTransform = AffineTransform.multiply(
-              AffineTransform.translation(newBoundingRect.x, newBoundingRect.y),
-              AffineTransform.scale(
-                newBoundingRect.width,
-                newBoundingRect.height,
-              ),
-            );
-
-            layerIndexPaths.forEach((layerIndex) => {
-              const originalLayer = Layers.access(pageSnapshot, layerIndex);
-
-              const layerTransform = AffineTransform.multiply(
-                ...Layers.accessPath(pageSnapshot, layerIndex)
-                  .slice(1, -1) // Remove the page and current layer
-                  .map((layer) =>
-                    AffineTransform.translation(layer.frame.x, layer.frame.y),
-                  )
-                  .reverse(),
-              );
-
-              const newLayer = Layers.access(
-                draft.sketch.pages[pageIndex],
-                layerIndex,
-              );
-
-              const min = AffineTransform.multiply(
-                layerTransform.invert(),
-                newTransform,
-                originalTransform,
-                layerTransform,
-              ).applyTo({
-                x: originalLayer.frame.x,
-                y: originalLayer.frame.y,
-              });
-
-              const max = AffineTransform.multiply(
-                layerTransform.invert(),
-                newTransform,
-                originalTransform,
-                layerTransform,
-              ).applyTo({
-                x: originalLayer.frame.x + originalLayer.frame.width,
-                y: originalLayer.frame.y + originalLayer.frame.height,
-              });
-
-              const newFrame = normalizeRect({
-                x: Math.round(min.x),
-                y: Math.round(min.y),
-                width: Math.round(max.x - min.x),
-                height: Math.round(max.y - min.y),
-              });
-
-              newLayer.frame.x = newFrame.x;
-              newLayer.frame.y = newFrame.y;
-              newLayer.frame.width = newFrame.width;
-              newLayer.frame.height = newFrame.height;
-            });
-
-            break;
-          }
-          case 'panning': {
-            const { previous, next } = interactionState;
-
-            const delta = {
-              x: next.x - previous.x,
-              y: next.y - previous.y,
-            };
-
-            const meta: EncodedPageMetadata = draft.sketch.user[currentPageId];
-
-            const parsed = Primitives.parsePoint(meta.scrollOrigin);
-
-            parsed.x += delta.x;
-            parsed.y += delta.y;
-
-            draft.sketch.user[currentPageId] = {
-              ...meta,
-              scrollOrigin: Primitives.stringifyPoint(parsed),
-            };
-
-            break;
-          }
-        }
-      });
     }
     case 'setTextColor':
     case 'setTextFontName':
